split DeployService class into multiple classes
