+++ modules/deployment/src/main/java/org/eclipse/xpanse/modules/deployment/SensitiveDataHandler.java
@@ -1,0 +1,77 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ * SPDX-FileCopyrightText: Huawei Inc.
+ *
+ */
+
+package org.eclipse.xpanse.modules.deployment;
+
+import jakarta.annotation.Resource;
+import java.util.Map;
+import java.util.Objects;
+import lombok.extern.slf4j.Slf4j;
+import org.eclipse.xpanse.modules.database.service.DeployServiceEntity;
+import org.eclipse.xpanse.modules.database.servicetemplate.ServiceTemplateEntity;
+import org.eclipse.xpanse.modules.models.servicetemplate.DeployVariable;
+import org.eclipse.xpanse.modules.models.servicetemplate.enums.SensitiveScope;
+import org.eclipse.xpanse.modules.security.common.AesUtil;
+import org.springframework.stereotype.Component;
+import org.springframework.util.CollectionUtils;
+
+/**
+ * Bean to manage service data which contains sensitive data.
+ */
+@Slf4j
+@Component
+public class SensitiveDataHandler {
+
+    @Resource
+    private AesUtil aesUtil;
+
+    /**
+     * Method to mask all sensitive data after deployment is completed.
+     *
+     * @param deployServiceEntity entity to be updated.
+     */
+    public void maskSensitiveFields(DeployServiceEntity deployServiceEntity) {
+        log.debug("masking sensitive input data after deployment");
+        if (Objects.nonNull(deployServiceEntity.getDeployRequest().getServiceRequestProperties())) {
+            for (DeployVariable deployVariable
+                    : deployServiceEntity.getDeployRequest().getOcl().getDeployment()
+                    .getVariables()) {
+                if (deployVariable.getSensitiveScope() != SensitiveScope.NONE
+                        && (deployServiceEntity.getDeployRequest().getServiceRequestProperties()
+                        .containsKey(deployVariable.getName()))) {
+                    deployServiceEntity.getDeployRequest().getServiceRequestProperties()
+                            .put(deployVariable.getName(), "********");
+
+                }
+            }
+        }
+    }
+
+    /**
+     * Method to encode all sensitive data in the request.
+     *
+     * @param serviceTemplate service template of the deployment request.
+     * @param serviceRequestProperties request properties sent by customer.
+     */
+    public void encodeDeployVariable(ServiceTemplateEntity serviceTemplate,
+                                      Map<String, Object> serviceRequestProperties) {
+        if (Objects.isNull(serviceTemplate.getOcl().getDeployment())
+                ||
+                CollectionUtils.isEmpty(serviceTemplate.getOcl().getDeployment().getVariables())
+                || Objects.isNull(serviceRequestProperties)) {
+            return;
+        }
+        serviceTemplate.getOcl().getDeployment().getVariables().forEach(variable -> {
+            if (Objects.nonNull(variable) && !SensitiveScope.NONE.toValue()
+                    .equals(variable.getSensitiveScope().toValue())
+                    && serviceRequestProperties.containsKey(variable.getName())) {
+                serviceRequestProperties.put(variable.getName(),
+                        aesUtil.encode(
+                                serviceRequestProperties.get(variable.getName()).toString()));
+            }
+        });
+    }
+}
