+++ modules/orchestrator/src/test/java/org/eclipse/xpanse/modules/orchestrator/deployment/DeployTaskTest.java
@@ -111,7 +111,7 @@ class DeployTaskTest {
                 + "csp=null, flavor=null, serviceHostingType=null, ocl=null, serviceRequestProperties=null), id=null),"
                 + " ocl=Ocl(category=null, version=null, name=null, serviceVersion=null, "
                 + "description=null, namespace=null, icon=null, cloudServiceProvider=null, "
-                + "deployment=null, flavors=null, billing=null, serviceHostingType=null), deployResourceHandler=null, servicePolicies=null)";
+                + "deployment=null, flavors=null, billing=null, serviceHostingType=null), deployResourceHandler=null, serviceTemplateId=null)";
         assertEquals(exceptedString, test.toString());
 
     }
